# examples/spring/srcgen/src/main/java/app/server/api/RoleApi.java
/**
 * NOTE: This class was generated by QSDL.
 */

package app.server.api;

import app.server.domain.*;
import app.server.model.*;

import org.springframework.http.*;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.*;


@RequestMapping("${server.base-path:/api/v1}")
@Controller
public interface RoleApi {

  /**
   * GET /projects/{project_id}/roles : List Roles.
   */
  @GetMapping(
      value = "/projects/{project_id}/roles",
      produces = {"application/json"}
  )
  default ResponseEntity<CursorPage<Role>> getRoles(@PathVariable("project_id") Long projectId, CursorPageable pageable) {
    return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
  }

  /**
   * POST /projects/{project_id}/roles : Create a Role.
   */
  @PostMapping(
      value = "/projects/{project_id}/roles",
      produces = {"application/json"},
      consumes = {"application/json"}
  )
  default ResponseEntity<Role> createRole(@PathVariable("project_id") Long projectId, @RequestBody Role body) {
    return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
  }

  /**
   * GET /projects/{project_id}/roles/{id} : Read the specified Role.
   */
  @GetMapping(
      value = "/projects/{project_id}/roles/{id}",
      produces = {"application/json"}
  )
  default ResponseEntity<Role> getRole(@PathVariable("project_id") Long projectId, @PathVariable("id") Long id) {
    return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
  }

  /**
   * PUT /projects/{project_id}/roles/{id} : Replace the specified Role.
   */
  @PutMapping(
      value = "/projects/{project_id}/roles/{id}",
      produces = {"application/json"},
      consumes = {"application/json"}
  )
  default ResponseEntity<Role> replaceRole(@PathVariable("project_id") Long projectId, @PathVariable("id") Long id, @RequestBody Role body) {
    return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
  }

  /**
   * PATCH /projects/{project_id}/roles/{id} : Update the specified Role.
   */
  @PatchMapping(
      value = "/projects/{project_id}/roles/{id}",
      produces = {"application/json"},
      consumes = {"application/json"}
  )
  default ResponseEntity<Role> updateRole(@PathVariable("project_id") Long projectId, @PathVariable("id") Long id, @RequestBody Role body) {
    return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
  }

  /**
   * DELETE /projects/{project_id}/roles/{id} : Delete the specified Role.
   */
  @DeleteMapping(
      value = "/projects/{project_id}/roles/{id}"
  )
  default ResponseEntity<Void> deleteRole(@PathVariable("project_id") Long projectId, @PathVariable("id") Long id) {
    return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
  }

}