# examples/spring/srcgen/src/main/java/app/server/api/TicketApi.java
/**
 * NOTE: This class was generated by QSDL.
 */

package app.server.api;

import app.server.domain.*;
import app.server.model.*;

import org.springframework.http.*;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.*;


@RequestMapping("${server.base-path:/api/v1}")
@Controller
public interface TicketApi {

  /**
   * GET /tickets : List Tickets.
   */
  @GetMapping(
      value = "/tickets",
      produces = {"application/json"}
  )
  default ResponseEntity<CursorPage<Ticket>> getTickets(CursorPageable pageable) {
    return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
  }

  /**
   * POST /tickets : Create a Ticket.
   */
  @PostMapping(
      value = "/tickets",
      produces = {"application/json"},
      consumes = {"application/json"}
  )
  default ResponseEntity<Ticket> createTicket(@RequestBody Ticket body) {
    return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
  }

  /**
   * GET /tickets/{id} : Read the specified Ticket.
   */
  @GetMapping(
      value = "/tickets/{id}",
      produces = {"application/json"}
  )
  default ResponseEntity<Ticket> getTicket(@PathVariable("id") Long id) {
    return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
  }

  /**
   * PUT /tickets/{id} : Replace the specified Ticket.
   */
  @PutMapping(
      value = "/tickets/{id}",
      produces = {"application/json"},
      consumes = {"application/json"}
  )
  default ResponseEntity<Ticket> replaceTicket(@PathVariable("id") Long id, @RequestBody Ticket body) {
    return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
  }

  /**
   * PATCH /tickets/{id} : Update the specified Ticket.
   */
  @PatchMapping(
      value = "/tickets/{id}",
      produces = {"application/json"},
      consumes = {"application/json"}
  )
  default ResponseEntity<Ticket> updateTicket(@PathVariable("id") Long id, @RequestBody Ticket body) {
    return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
  }

  /**
   * DELETE /tickets/{id} : Delete the specified Ticket.
   */
  @DeleteMapping(
      value = "/tickets/{id}"
  )
  default ResponseEntity<Void> deleteTicket(@PathVariable("id") Long id) {
    return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
  }

}