# examples/spring/srcgen/src/main/java/app/server/controller/ProjectController.java
/**
 * NOTE: This class was generated by QSDL.
 */

package app.server.controller;

import app.server.api.ProjectApi;
import app.server.domain.*;
import app.server.model.*;
import app.server.service.ProjectService;
import app.server.util.Validator;

import lombok.AllArgsConstructor;
import org.springframework.http.*;
import org.springframework.stereotype.Controller;


@Controller
@AllArgsConstructor
public class ProjectController extends BaseController implements ProjectApi {

  private final ProjectService projectService;

  /**
   * {@inheritDoc}.
   */
  @Override
  public ResponseEntity<CursorPage<Project>> getProjects(CursorPageable pageable) {
    CursorPage<Project> response = projectService.getProjects(pageable, super.getContext());
    return new ResponseEntity<>(response, HttpStatus.OK);
  }

  /**
   * {@inheritDoc}.
   */
  @Override
  public ResponseEntity<Project> createProject(Project body) {
    Validator.validate(body);
    Project response = projectService.createProject(body, super.getContext());
    return new ResponseEntity<>(response, HttpStatus.OK);
  }

  /**
   * {@inheritDoc}.
   */
  @Override
  public ResponseEntity<Project> getProject(Long id) {
    Project response = projectService.getProject(id, super.getContext());
    return new ResponseEntity<>(response, HttpStatus.OK);
  }

  /**
   * {@inheritDoc}.
   */
  @Override
  public ResponseEntity<Project> replaceProject(Long id, Project body) {
    Validator.validate(body);
    Project response = projectService.replaceProject(id, body, super.getContext());
    return new ResponseEntity<>(response, HttpStatus.OK);
  }

  /**
   * {@inheritDoc}.
   */
  @Override
  public ResponseEntity<Project> updateProject(Long id, Project body) {
    Validator.validateExRequired(body);
    Project response = projectService.updateProject(id, body, super.getContext());
    return new ResponseEntity<>(response, HttpStatus.OK);
  }

  /**
   * {@inheritDoc}.
   */
  @Override
  public ResponseEntity<Void> deleteProject(Long id) {
    projectService.deleteProject(id, super.getContext());
    return new ResponseEntity<>(HttpStatus.OK);
  }

}