# examples/spring/srcgen/src/main/java/app/server/domain/Project.java
/**
 * NOTE: This class was generated by QSDL.
 */

package app.server.domain;

import app.server.model.AbstractClass;

import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.databind.node.ObjectNode;

import jakarta.validation.constraints.*;

import java.time.*;

public class Project extends AbstractClass {

  @Min(0)
  @Max(Long.MAX_VALUE)
  @JsonProperty(value = "id", required = true, access = JsonProperty.Access.READ_ONLY)
  public Long id;

  @NotNull
  @Size(min = 0, max = 255)
  @JsonProperty(value = "name", required = true)
  public String name;

  @Size(min = 0, max = 255)
  @JsonProperty(value = "description")
  public String description;

  @Size(min = 0, max = 255)
  @JsonProperty(value = "creation_by", access = JsonProperty.Access.READ_ONLY)
  public String creationBy;

  @JsonProperty(value = "creation_date", access = JsonProperty.Access.READ_ONLY)
  public LocalDate creationDate;

  @Size(min = 0, max = 255)
  @JsonProperty(value = "last_update_by", access = JsonProperty.Access.READ_ONLY)
  public String lastUpdateBy;

  @JsonProperty(value = "last_update_date", access = JsonProperty.Access.READ_ONLY)
  public OffsetDateTime lastUpdateDate;

  @JsonProperty(value = "meta_inf")
  public ObjectNode metaInf;

  @JsonProperty(value = "archive", access = JsonProperty.Access.WRITE_ONLY)
  public Boolean archive;

}