# examples/spring/srcgen/src/main/java/app/server/domain/Role.java
/**
 * NOTE: This class was generated by QSDL.
 */

package app.server.domain;

import app.server.model.AbstractClass;

import com.fasterxml.jackson.annotation.JsonProperty;

import jakarta.validation.constraints.*;

public class Role extends AbstractClass {

  @Min(0)
  @Max(Long.MAX_VALUE)
  @JsonProperty(value = "id", required = true, access = JsonProperty.Access.READ_ONLY)
  public Long id;

  @NotNull
  @Size(min = 0, max = 255)
  @JsonProperty(value = "name", required = true)
  public String name;

}